requirements.txt
biopython==1.78
pandas
numpy
